install.sh now installs piHeatingHub

+++ scripts/piHeatingHubInstall.sh
@@ -2,54 +2,75 @@
 
 ARG1=$1
 ARG2=$2
+ARG3=$3
+ARG4=$4
+ARG5=$5
+ARG6=$6
+ARG7=$7
+ARG8=$8
 
 ########## check for arguments
-if [ ! $ARG1 ] || [ ! $ARG2]; then
-	echo -e "\n\n Error: \n This script must be run with arguments, \n or rather started from main script \n Exiting ..."
+if [[ ! $ARG1 ]] || [[ ! $ARG2 ]] || [[ ! $ARG3 ]] || [[ ! $ARG4 ]] || [[ ! $ARG5 ]] || [[ ! $ARG6 ]] || [[ ! $ARG7 ]]; then
+	echo -e "\n\n Error: \n This script must be run with arguments, \n or rather started from main script 'install.sh' \n Exiting ..."
 	exit 1
 else
 	scriptDir=$ARG1
 	installDir=$ARG2
+	DB_USERNAME=$ARG3
+	DB_PASSWORD=$ARG4
+	DB_SERVER=$ARG5
+	DB_NAME=$ARG6
+	simulateMessage=$ARG7
+	simulate=$ARG8
 fi
 
 ########## check for root
 if [[ `whoami` != "root" ]]; then
   	echo -e "\n\n Error: \n Script must be run as root  \n Exiting ..."
-  	exit 1
+  	if [ $simulate ]; then
+		echo "$simulateMessage"
+	else
+  		exit 1
+  	fi
 fi
 
 ########## installation
 ########## install binaries and web
+echo -e "\n\n Installing piHeatingHub ... \n ----------"
 if [ ! -f "$installDir/piHeatingHub/README.md" ]; then
-	if [ -d "$installDir/piHeatingHub" ]; then
-		echo -e " Deleting old directory ..."
-    	rm -rf "$installDir/piHeatingHub"
-  	fi
-
-	echo -e " Copying binaries ..."
-  	cp -rf "$scriptDir/piHeating/piHeatingHub" "$installDir/piHeatingHub"
-  	echo -e " Moving web ..."
-  	mv "$installDir/piHeatingHub/www" "/var/www/piHeatingHub"
-  
-  	echo -e " Creating data directories ..."
-  	if [ ! -d "/var/www/piHeatingHub/data" ]; then
-    	mkdir "/var/www/piHeatingHub/data"
-  	fi
-  	mkdir "$installDir/piHeatingHub/data"
-  	
-  	echo -e " Setting permissions ..."
-  	chown -R pi:www-data "$installDir/piHeatingHub"
-  	chmod -R 750 "$installDir/piHeatingHub"
-  	chown -R pi:www-data "$installDir/piHeatingHub/data"
-  	chmod -R 775 "$installDir/piHeatingHub/data"
-  
-  	chown -R pi:www-data "/var/www/piHeatingHub"
-  	chmod -R 755 "/var/www/piHeatingHub"
-  	chmod -R 775 "/var/www/piHeatingHub/images"
-
-	if [ ! -f "$installDir/piHeatingHub/README.md" ]; then
-		echo " Error: \n piHeatingHub installation failed\n"
-		exit 1
+	if [ $simulate ]; then
+		echo "$simulateMessage"
+	else
+		if [ -d "$installDir/piHeatingHub" ]; then
+			echo -e " Deleting old directory ..."
+	    	rm -rf "$installDir/piHeatingHub"
+	  	fi
+	
+		echo -e " Copying binaries ..."
+	  	cp -rf "$scriptDir/piHeating/piHeatingHub" "$installDir/piHeatingHub"
+	  	echo -e " Moving web ..."
+	  	mv "$installDir/piHeatingHub/www" "/var/www/piHeatingHub"
+	  
+	  	echo -e " Creating data directories ..."
+	  	if [ ! -d "/var/www/piHeatingHub/data" ]; then
+	    	mkdir "/var/www/piHeatingHub/data"
+	  	fi
+	  	mkdir "$installDir/piHeatingHub/data"
+	  	
+	  	echo -e " Setting permissions ..."
+	  	chown -R pi:www-data "$installDir/piHeatingHub"
+	  	chmod -R 750 "$installDir/piHeatingHub"
+	  	chown -R pi:www-data "$installDir/piHeatingHub/data"
+	  	chmod -R 775 "$installDir/piHeatingHub/data"
+	  
+	  	chown -R pi:www-data "/var/www/piHeatingHub"
+	  	chmod -R 755 "/var/www/piHeatingHub"
+	  	chmod -R 775 "/var/www/piHeatingHub/images"
+	
+		if [ ! -f "$installDir/piHeatingHub/README.md" ]; then
+			echo " Error: \n piHeatingHub installation failed\n"
+			exit 1
+		fi
 	fi
 else
   	echo " piHeatingHub is already installed"
@@ -57,26 +78,37 @@ fi
 
 ########## create cron job
 echo -e " Creating cron jobs ..."
-if [ ! -f "/etc/cron.d/piHeating" ]; then
-    cat > /etc/cron.d/piHeating <<CRON
+if [ $simulate ]; then
+  	echo "$simulateMessage"
+else
+	if [ ! -f "/etc/cron.d/piHeating" ]; then
+	    cat > /etc/cron.d/piHeating <<CRON
 MAILTO=""
 * * * * * pi /bin/bash $installDir/piHeatingHub/cron/piHeatingHubWrapper.sh >> /dev/null 2>&1
 CRON
+	fi
 fi
 		
 ########## configure apache
 echo -e " Setting Apache listen port 8080 ..."
-if grep -Fxq 'Listen 8080' /etc/apache2/ports.conf; then
+if grep -Fxq 'Listen 8080' /etc/apache2/ports.conf >> /dev/null 2>&1; then
 	printf "   Apache already listening on port 8080 \n"
 else
-	cat >> /etc/apache2/ports.conf <<PORTS
+	if [ $simulate ]; then
+  		echo "$simulateMessage"
+  	else
+		cat >> /etc/apache2/ports.conf <<PORTS
 Listen 8080
 PORTS
+	fi
 fi
 
 ########## add apache site
 echo -e " Adding site ..."
-cat > /etc/apache2/sites-available/piHeatingHub.conf <<VHOST
+if [ $simulate ]; then
+  	echo "$simulateMessage"
+else
+	cat > /etc/apache2/sites-available/piHeatingHub.conf <<VHOST
 <VirtualHost *:8080>
     ServerAdmin webmaster@localhost
     DocumentRoot /var/www/piHeatingHub/
@@ -91,18 +123,29 @@ cat > /etc/apache2/sites-available/piHeatingHub.conf <<VHOST
     CustomLog \${APACHE_LOG_DIR}/access.log combined
 </VirtualHost>
 VHOST
+fi
 		
 echo -e " Enabling site ..."
-a2ensite piHeatingHub.conf
+if [ $simulate ]; then
+  	echo "$simulateMessage"
+else
+	a2ensite piHeatingHub.conf
+fi
 
 ########## database setup
-echo -e "\n\n Setting up database ... \n ----------"
-echo -e " Please enter the MariaDB 'root' password : "
-read -s ROOT_PASSWORD
-
-echo -e "     Creating database: $DB_NAME \n at host: $DB_SERVER \n Setting username: $DB_USER \n with password: $DB_PASSWORD"
-mysql -uroot -p$ROOT_PASSWORD<< DATABASE
-
+echo -e " Setting up database ..."
+if [ -d /var/lib/mysql/piHeatingDB ]; then
+	echo " Database already exists"
+else
+	if [ $simulate ]; then
+	  	echo "$simulateMessage"
+	else
+		echo -e " Please enter the MariaDB 'root' password : "
+		read -s ROOT_PASSWORD
+		
+		echo -e "     Creating database: $DB_NAME \n     at host: $DB_SERVER \n     Setting username: $DB_USER \n     with password: $DB_PASSWORD"
+		mysql -uroot -p$ROOT_PASSWORD<< DATABASE
+	
 CREATE DATABASE $DB_NAME CHARACTER SET = utf8;
 
 CREATE USER '$DB_USERNAME'@'$DB_SERVER';
@@ -180,14 +223,19 @@ CREATE TABLE IF NOT EXISTS sched_network (  sn_id         int(11)       NOT NULL
                                             network_id    int(11)       DEFAULT NULL,
                                             test          tinyint(1)    DEFAULT NULL );
 
-DATABASE	
-		
-echo -e "\n\n Saving database credentials ... \n ----------"
-cat > $installDir/piHeatingHub/config/config.ini <<CONFIG
+DATABASE
+	fi
+			
+	echo -e " Saving database credentials ..."
+	if [ $simulate ]; then
+	  	echo "$simulateMessage"
+	else
+		cat > $installDir/piHeatingHub/config/config.ini <<CONFIG
 [db]
 server = $DB_SERVER
 user = $DB_USERNAME
 password = $DB_PASSWORD
 database = $DB_NAME
 CONFIG
-
+	fi
+fi
